[1067] Fail gracefully when trying to encrypt a file > 2GB.
Note that supporting file encryption for files > 2GB is not doable
without breaking backwards compatability, since we encrypt the length
of the data first using 32 bits. Fixable, but I wondier if it's worth
the effort. In any case, I've changed the file length type to be 64 bits,
(a) to detect this gracefully, and (b) towards supporting password
databases that may exceed 2GB, e.g., if/when we have attachments.
(Note that the 2G limit is inherent on a single field, not the file...)

+++ src/core/PWScore.cpp
@@ -935,7 +935,7 @@ static void ManageIncBackupFiles(const stringT &cs_filenamebase,
 
   stringT cs_filenamemask(cs_filenamebase);
   vector<stringT> files;
-  vector<int> file_nums;
+  vector<unsigned int> file_nums;
 
   cs_filenamemask += _T("_???.ibak");
 
@@ -961,7 +961,7 @@ static void ManageIncBackupFiles(const stringT &cs_filenamebase,
   sort(file_nums.begin(), file_nums.end());
 
   // nnn is the number of the file in the returned value: cs_filebasename_nnn
-  int nnn = file_nums.back();
+  size_t nnn = file_nums.back();
   nnn++;
   if (nnn > 999) {
     // as long as there's a _999 file, we set n starting from 001
